Split out CLI, show actual bundle names
